restructuring elpi generation
